app/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
numpy==1.24.3
opencv-python==4.8.0.76
pillow==10.0.0
python-multipart==0.0.6
